Estado de blackjack Limpar mão Add histórico Apostar Obter nome Criar jogador

+++ Jogador.hs
@@ -7,5 +7,5 @@ import Carta
 data Jogador = Jogador
     { nome :: String,
       mao :: Mao,
-      historico :: [Int]
+      historico :: [Float]
     } deriving Show
